+++ contracts/brane_auction/src/msgs.rs
@@ -19,9 +19,9 @@ pub struct InstantiateMsg {
     /// First submission for the first NFT auction of the collection
     pub first_submission: SubmissionInfo,
     ///Mint cost
-    /// testnet: 50_000_000u128
-    /// mainnet: 5_000_000u128
-    pub mint_cost: u128,
+    /// testnet: 50_000_000u64
+    /// mainnet: 5_000_000u64
+    pub mint_cost: u64,
 }
 
 #[cw_serde]
